Added WarlordsPlayerDropWeaponEvent Added WEAPON_DROP_RATE upgrade implementation

+++ src/main/java/com/ebicep/warlords/game/option/wavedefense/mobs/Mob.java
@@ -10,7 +10,7 @@ import java.util.UUID;
 
 public interface Mob {
 
-    int dropRate();
+    double dropRate();
 
     int commonDropChance();
