+++ src/main/java/com/ebicep/warlords/game/option/wavedefense/mobs/mobtypes/BasicMob.java
@@ -5,7 +5,7 @@ import com.ebicep.warlords.game.option.wavedefense.mobs.Mob;
 public interface BasicMob extends Mob {
 
     @Override
-    default int dropRate() {
+    default double dropRate() {
         return 1;
     }
